+++ ChangeLog.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 
 - added support for getting db log entries via `GET /_admin/log/entries` (ArangoDB v3.8)
+- added support for index estimates (ArangoDB v3.8)
+- added support for ArangoSearch `AQL`, `Pipeline`, `Stopwords`, `GeoJSON`, `GeoPoint` analyzers
+- fixed active failover behavior for the asynchronous driver 
+- deprecated `ArangoIterable` methods in favour of Java 8 Stream equivalents
 
 ## [6.10.0] - 2021-03-27